Refactor CalculatorStoreProvider and update imports for layout components
